fix: App with new OfferRecomended component
